replace README file with our docker instructions
We will run into merge conflicts this way if we want to integrate changes from
upstream. On the other side I didn't want to keep the old instructions.

I see no value in having the instructions for the pm2 process manager. Docker
and systemd are already capable of restarting a process.