Catch uncaught exception when node disconnects
The getStats function handles the error of all RPC calls except one.
Since all calls for the stats are run in parallel, this always lead to
an uncaught exception. The fix simple catches this and forward it to the
callback function as for every other call as well.

Related to: #10